requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
PyMuPDF==1.24.2
opencv-python==4.9.0.80
pyodbc==5.1.0
pillow==10.3.0
easyocr==1.7.1
datetime